+++ src/applications/dependents/686c-674/config/utilities/index.js
@@ -19,6 +19,12 @@ import {
   spouseEvidence,
   childEvidence,
   showDupeModalIfEnabled,
+  isAddingDependents,
+  isRemovingDependents,
+  showV3Picklist,
+  noV3Picklist,
+  hasAwardedDependents,
+  isVisiblePicklistPage,
 } from './data';
 
 export {
@@ -42,4 +48,10 @@ export {
   spouseEvidence,
   childEvidence,
   showDupeModalIfEnabled,
+  isAddingDependents,
+  isRemovingDependents,
+  showV3Picklist,
+  noV3Picklist,
+  hasAwardedDependents,
+  isVisiblePicklistPage,
 };
