+++ src/applications/dependents/686c-674/containers/App.jsx
@@ -51,6 +51,7 @@ function App({
   useFormFeatureToggleSync([
     'vaDependentsNetWorthAndPension',
     'vaDependentsDuplicateModals',
+    'vaDependentsV3',
   ]);
   const dependentsModuleEnabled = useToggleValue(
     TOGGLE_NAMES.dependentsModuleEnabled,
