+++ src/applications/dependents/shared/utils/index.js
@@ -1,6 +1,8 @@
-import { parse, isValid, format } from 'date-fns';
-
 import { srSubstitute } from 'platform/forms-system/src/js/utilities/ui/mask-string';
+
+import { getFormatedDate, calculateAge } from './dates';
+
+export { getFormatedDate, calculateAge };
 
 /**
  * Return formatted full name from name object
@@ -14,22 +16,6 @@ import { srSubstitute } from 'platform/forms-system/src/js/utilities/ui/mask-str
 export const getFullName = ({ first, middle, last, suffix } = {}) =>
   [first || '', middle || '', last || ''].filter(Boolean).join(' ') +
   (suffix ? `, ${suffix}` : '');
-
-/**
- *
- * @param {String} date - date string that matches the `startFormat`
- * @param {String} startFormat - input date string format using Date-fn patterns
- * @param {String} endFormat - output date string format using Date-fn patterns
- * @returns {String} - formatted date string, or 'Unknown date' if invalid
- */
-export const getFormatedDate = (
-  date,
-  startFormat = 'yyyy-MM-dd',
-  endFormat = 'MMMM d, yyyy',
-) => {
-  const dobObj = parse(date, startFormat, new Date());
-  return isValid(dobObj) ? format(dobObj, endFormat) : date || 'Unknown date';
-};
 
 /**
  * Separate each number so the screen reader reads "number ending with 1 2 3 4"
